Refactor ft_time.c for improved time handling

+++ philo/ft_time.c
@@ -6,7 +6,7 @@
 /*   By: asodor <asodor@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/29 06:25:12 by asodor            #+#    #+#             */
-/*   Updated: 2024/12/22 18:00:38 by asodor           ###   ########.fr       */
+/*   Updated: 2024/12/23 16:05:52 by asodor           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 unsigned int	ft_get_time(void)
 {
-	struct timeval		time;
-	static unsigned int	start;
+	struct timeval			time;
+	static unsigned long	start;
 
 	if (!start)
 	{
@@ -24,8 +24,7 @@ unsigned int	ft_get_time(void)
 		return (0);
 	}
 	gettimeofday(&time, NULL);
-	return ((unsigned long)(time.tv_sec * 1000) + (time.tv_usec / 1000)
-		- start);
+	return ((unsigned int)(time.tv_sec * 1000) + (time.tv_usec / 1000) - start);
 }
 
 int	ft_usleep(unsigned long time)
